% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_plot.R
\name{generate_density_plot}
\alias{generate_density_plot}
\title{Generate density plot of multiverse analysis results}
\usage{
generate_density_plot(data, outcome_var, outcome_var_label)
}
\arguments{
\item{data}{A dataframe containing results of the multiverse analysis.
Must include the outcome variable and a binary column \code{significant}.}

\item{outcome_var}{Character. The name of the outcome variable to plot.}

\item{outcome_var_label}{Character. Label for the outcome variable (x-axis).}
}
\value{
A ggplot object representing the density plot.
}
\description{
This function creates a density plot comparing the distribution of all
estimated outcomes with the distribution of statistically significant outcomes
in a multiverse analysis.
}
\examples{
prep <- prep_data("hurricane")
df <- prep[[1]]
generate_density_plot(df, "edif", "Excess Fatalities")

}
