% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_strips.R
\name{generate_heatmap_strips}
\alias{generate_heatmap_strips}
\title{Generate heatmap strips for multiverse outcomes}
\usage{
generate_heatmap_strips(data, outcome_var, strip_vars, variable_labels)
}
\arguments{
\item{data}{A dataframe with results of the multiverse analysis.}

\item{outcome_var}{Character. Name of the outcome variable.}

\item{strip_vars}{Character vector. Names of decision variables to display.}

\item{variable_labels}{Named vector. Display labels for the decision variables.}
}
\value{
A list of ggplot objects representing heatmap strips.
}
\description{
Creates a set of heatmap strip plots that visualize how outcome estimates
vary depending on analytical decision variables in a multiverse analysis.
}
\examples{
prep <- prep_data("hurricane")
df <- prep[[1]]
strip_vars <- prep[[4]]
variable_labels <- prep[[5]]
generate_heatmap_strips(df, "edif", strip_vars, variable_labels)

}
