% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Prepare data for multiverse analysis plots}
\usage{
prep_data(use_case)
}
\arguments{
\item{use_case}{Either "hurricane" or "beauty".}
}
\value{
A list with data and metadata
}
\description{
Prepare data for multiverse analysis plots
}
